fix(api): move price history schemas into models
